Create README.md
People are complaining about Airlines all the time. I wanted to see which are the complaints, which are not. With Twitter text data, I defined each word as pos or neg by matching them with online lexicons and generated a sentiment score for each tweet using the function: #pos-#neg. I improved the accuracy using maximum entropy and screened out 43 non-complaint tweets out of thousands of tweets.